+++ jupytab-server/jupytab_server/jupytab.py
@@ -11,9 +11,9 @@ from configparser import ConfigParser, NoSectionError, NoOptionError
 from tornado.ioloop import IOLoop
 from tornado.web import StaticFileHandler, Application
 
-from jupytab.kernel_executor import KernelExecutor
-from jupytab.jupytab_api import RestartHandler, APIHandler, ReverseProxyHandler, root, api_kernel, \
-    access_kernel, restart_kernel
+from jupytab_server.kernel_executor import KernelExecutor
+from jupytab_server.jupytab_api import RestartHandler, APIHandler, ReverseProxyHandler, root, \
+    api_kernel, access_kernel, restart_kernel
 
 logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)